requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
lxml==4.3.0
Pillow==5.4.1
pip==18.1
psycopg2==2.7.6.1
pytz==2018.7
requests==2.21.0
setuptools==40.6.3
urllib3==1.24.1
virtualenv==16.2.0
wheel==0.32.3
